requirements.txt
openai
python-dotenv
requests
PyMuPDF
watchdog
tenacity
Flask
gunicorn
desktop-notifier
google-generativeai
chromadb==0.5.0
numpy==1.26.4
pydantic
instructor
llama-index
llama-index-llms-openai
llama-index-embeddings-google
datasets
